dinamic path alternative 1

+++ src/app/core/components/settings/settings.component.ts
@@ -1,6 +1,6 @@
-import { Component, inject } from '@angular/core';
+import {Component, inject, OnInit} from "@angular/core";
 import {StandardTopBarComponent} from "../top-bars/standard-top-bar/standard-top-bar.component";
-import {RouterOutlet} from "@angular/router";
+import {ActivatedRoute, Router, RouterOutlet} from "@angular/router";
 import {NgClass} from "@angular/common";
 import {ActiveWindowService} from "../../services/active-window.service";
 
@@ -15,17 +15,32 @@ import {ActiveWindowService} from "../../services/active-window.service";
     templateUrl: "./settings.component.html",
     styleUrl: "./settings.component.scss"
 })
-export class SettingsComponent {
-    private activeWindowService = inject(ActiveWindowService);
-
+export class SettingsComponent implements OnInit {
     settingComponent = "general"
+    isNative: Boolean = false;
+    private activeWindowService = inject(ActiveWindowService);
+    private activatedRoute = inject(ActivatedRoute);
+    private router = inject(Router);
+
+
+    ngOnInit() {
+        this.activatedRoute.data.subscribe(data => {
+            this.isNative = data["native"] ?? false;
+            this.settingComponent = data["component"] ?? "general";
+        });
+    }
+
 
     setComponent(componentName: string) {
         this.settingComponent = componentName;
     }
 
     closeWindow() {
-        this.activeWindowService.removeWindow("settings");
+        if (this.isNative) {
+            this.router.navigate(["/"]).then();
+        } else {
+            this.activeWindowService.removeWindow("settings");
+        }
     }
 
 }
